리소스 수정, 피드 리팩토링, 간격 수정
