+++ core/data/src/main/res/drawable/bar_end_image.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,12m-12,0a12,12 0,1 1,24 0a12,12 0,1 1,-24 0"
+      android:fillColor="#3E4049"/>
+  <path
+      android:pathData="M12.823,5.882C12.823,6.922 12.455,7.765 12,7.765C11.545,7.765 11.176,6.922 11.176,5.882C11.176,4.843 11.545,4 12,4C12.455,4 12.823,4.843 12.823,5.882ZM12.299,9.144C12.183,8.883 11.813,8.882 11.695,9.141L10.946,10.793C10.912,10.866 10.854,10.924 10.78,10.957L9.138,11.699C8.879,11.816 8.879,12.185 9.138,12.302L10.78,13.043C10.854,13.076 10.912,13.135 10.946,13.208L11.695,14.859C11.813,15.119 12.183,15.118 12.299,14.857L13.031,13.21C13.064,13.136 13.123,13.076 13.198,13.043L14.854,12.302C15.115,12.186 15.115,11.815 14.854,11.698L13.198,10.958C13.123,10.925 13.064,10.865 13.031,10.79L12.299,9.144ZM14.995,9.005C15.317,9.326 16.174,8.991 16.909,8.256C17.644,7.521 17.979,6.664 17.657,6.343C17.336,6.021 16.479,6.356 15.744,7.091C15.009,7.827 14.674,8.683 14.995,9.005ZM18.118,12.824C17.078,12.824 16.235,12.455 16.235,12C16.235,11.545 17.078,11.177 18.118,11.177C19.157,11.177 20,11.545 20,12C20,12.455 19.157,12.824 18.118,12.824ZM14.994,14.994C14.673,15.316 15.008,16.173 15.743,16.908C16.478,17.643 17.335,17.978 17.656,17.656C17.978,17.335 17.643,16.478 16.908,15.743C16.173,15.008 15.316,14.673 14.994,14.994ZM12.823,18.118C12.823,19.157 12.455,20 12,20C11.545,20 11.176,19.157 11.176,18.118C11.176,17.078 11.545,16.235 12,16.235C12.455,16.235 12.823,17.078 12.823,18.118ZM6.343,17.657C6.665,17.979 7.522,17.643 8.257,16.908C8.992,16.173 9.327,15.316 9.005,14.995C8.684,14.673 7.827,15.008 7.092,15.743C6.357,16.479 6.022,17.335 6.343,17.657ZM5.882,12.824C4.843,12.824 4,12.455 4,12C4,11.545 4.843,11.177 5.882,11.177C6.922,11.177 7.765,11.545 7.765,12C7.765,12.455 6.922,12.824 5.882,12.824ZM6.343,6.343C6.022,6.665 6.357,7.521 7.092,8.257C7.827,8.992 8.684,9.327 9.005,9.005C9.327,8.684 8.992,7.827 8.257,7.092C7.521,6.357 6.665,6.022 6.343,6.343Z"
+      android:fillColor="#ffffff"
+      android:fillType="evenOdd"/>
+</vector>
