+++ core/data/src/main/res/drawable/bar_start_image.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,12m-12,0a12,12 0,1 1,24 0a12,12 0,1 1,-24 0"
+      android:fillColor="#3E4049"/>
+  <path
+      android:pathData="M12.723,5.627C12.443,4.997 11.549,4.993 11.264,5.621L9.618,9.246C9.538,9.422 9.396,9.564 9.219,9.644L5.615,11.271C4.987,11.554 4.987,12.446 5.615,12.729L9.219,14.356C9.396,14.436 9.538,14.578 9.618,14.754L11.264,18.379C11.549,19.007 12.443,19.003 12.723,18.373L14.329,14.76C14.409,14.58 14.553,14.435 14.733,14.355L18.367,12.73C18.998,12.448 18.998,11.552 18.367,11.27L14.733,9.645C14.553,9.565 14.409,9.42 14.329,9.24L12.723,5.627Z"
+      android:fillColor="#ffffff"/>
+</vector>
