+++ core/data/src/main/res/drawable/location.xml
@@ -1,0 +1,17 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="10dp"
+    android:height="10dp"
+    android:viewportWidth="10"
+    android:viewportHeight="10">
+  <group>
+    <clip-path
+        android:pathData="M0,0h10v10h-10z"/>
+    <path
+        android:pathData="M5,9.583C5,9.583 8.75,7.083 8.75,4.167C8.75,3.172 8.355,2.218 7.652,1.515C6.948,0.812 5.995,0.417 5,0.417C4.005,0.417 3.052,0.812 2.348,1.515C1.645,2.218 1.25,3.172 1.25,4.167C1.25,7.083 5,9.583 5,9.583ZM6.667,4.166C6.667,5.087 5.921,5.833 5,5.833C4.08,5.833 3.333,5.087 3.333,4.166C3.333,3.246 4.08,2.5 5,2.5C5.921,2.5 6.667,3.246 6.667,4.166Z"
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"/>
+    <path
+        android:pathData="M5,9.583L4.769,9.93C4.909,10.023 5.091,10.023 5.231,9.93L5,9.583ZM7.652,1.515L7.357,1.809V1.809L7.652,1.515ZM2.348,1.515L2.643,1.809V1.809L2.348,1.515ZM8.333,4.167C8.333,5.457 7.493,6.71 6.573,7.682C6.121,8.159 5.667,8.55 5.326,8.821C5.156,8.956 5.015,9.062 4.917,9.132C4.868,9.168 4.83,9.194 4.805,9.212C4.792,9.221 4.783,9.227 4.777,9.231C4.774,9.233 4.771,9.235 4.77,9.236C4.769,9.236 4.769,9.236 4.769,9.237C4.769,9.237 4.769,9.237 4.769,9.237C4.769,9.237 4.769,9.237 4.769,9.237C4.769,9.237 4.769,9.236 5,9.583C5.231,9.93 5.231,9.93 5.231,9.93C5.231,9.93 5.232,9.93 5.232,9.929C5.232,9.929 5.232,9.929 5.233,9.929C5.233,9.928 5.234,9.928 5.235,9.927C5.238,9.925 5.241,9.923 5.245,9.921C5.253,9.915 5.265,9.907 5.279,9.897C5.309,9.876 5.351,9.847 5.405,9.808C5.512,9.73 5.664,9.617 5.845,9.473C6.208,9.184 6.692,8.768 7.177,8.255C8.132,7.248 9.167,5.792 9.167,4.167H8.333ZM7.357,1.809C7.982,2.435 8.333,3.282 8.333,4.167H9.167C9.167,3.061 8.728,2.002 7.946,1.22L7.357,1.809ZM5,0.833C5.884,0.833 6.732,1.184 7.357,1.809L7.946,1.22C7.165,0.439 6.105,-0 5,-0V0.833ZM2.643,1.809C3.268,1.184 4.116,0.833 5,0.833V-0C3.895,-0 2.835,0.439 2.054,1.22L2.643,1.809ZM1.667,4.167C1.667,3.282 2.018,2.435 2.643,1.809L2.054,1.22C1.272,2.002 0.833,3.061 0.833,4.167H1.667ZM5,9.583C5.231,9.236 5.231,9.237 5.231,9.237C5.231,9.237 5.231,9.237 5.231,9.237C5.231,9.237 5.231,9.237 5.231,9.237C5.231,9.236 5.231,9.236 5.23,9.236C5.229,9.235 5.226,9.233 5.223,9.231C5.217,9.227 5.208,9.221 5.195,9.212C5.17,9.194 5.132,9.168 5.083,9.132C4.985,9.062 4.844,8.956 4.674,8.821C4.333,8.55 3.879,8.159 3.427,7.682C2.507,6.71 1.667,5.457 1.667,4.167H0.833C0.833,5.792 1.868,7.248 2.823,8.255C3.308,8.768 3.792,9.184 4.155,9.473C4.336,9.617 4.488,9.73 4.595,9.808C4.649,9.847 4.691,9.876 4.721,9.897C4.735,9.907 4.747,9.915 4.755,9.921C4.759,9.923 4.762,9.925 4.765,9.927C4.766,9.928 4.767,9.928 4.767,9.929C4.768,9.929 4.768,9.929 4.768,9.929C4.768,9.93 4.769,9.93 4.769,9.93C4.769,9.93 4.769,9.93 5,9.583ZM5,6.25C6.151,6.25 7.083,5.317 7.083,4.166H6.25C6.25,4.857 5.69,5.416 5,5.416V6.25ZM2.917,4.166C2.917,5.317 3.85,6.25 5,6.25V5.416C4.31,5.416 3.75,4.857 3.75,4.166H2.917ZM5,2.083C3.85,2.083 2.917,3.016 2.917,4.166H3.75C3.75,3.476 4.31,2.916 5,2.916V2.083ZM7.083,4.166C7.083,3.016 6.151,2.083 5,2.083V2.916C5.69,2.916 6.25,3.476 6.25,4.166H7.083Z"
+        android:fillColor="#ffffff"/>
+  </group>
+</vector>
